chore(codersdk): move all wireguard functions to `workspacesdk`
Currently, importing `codersdk` just to interact with the API requires
importing tailscale, which causes builds to fail unless manually using
our fork.
