+++ coderd/prometheusmetrics/insights/metricscollector_test.go
@@ -26,6 +26,7 @@ import (
 	"github.com/coder/coder/v2/coderd/workspaceapps"
 	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/codersdk/agentsdk"
+	"github.com/coder/coder/v2/codersdk/workspacesdk"
 	"github.com/coder/coder/v2/provisioner/echo"
 	"github.com/coder/coder/v2/provisionersdk/proto"
 	"github.com/coder/coder/v2/testutil"
@@ -129,9 +130,10 @@ func TestCollectInsights(t *testing.T) {
 	defer closeFunc()
 
 	// Connect to the agent to generate usage/latency stats.
-	conn, err := client.DialWorkspaceAgent(ctx, resources[0].Agents[0].ID, &codersdk.DialWorkspaceAgentOptions{
-		Logger: logger.Named("client"),
-	})
+	conn, err := workspacesdk.NewWorkspaceClient(client).
+		DialWorkspaceAgent(ctx, resources[0].Agents[0].ID, &workspacesdk.DialWorkspaceAgentOptions{
+			Logger: logger.Named("client"),
+		})
 	require.NoError(t, err)
 	defer conn.Close()
