+++ coderd/tailnet.go
@@ -23,7 +23,7 @@ import (
 	"cdr.dev/slog"
 	"github.com/coder/coder/v2/coderd/tracing"
 	"github.com/coder/coder/v2/coderd/workspaceapps"
-	"github.com/coder/coder/v2/codersdk"
+	"github.com/coder/coder/v2/codersdk/workspacesdk"
 	"github.com/coder/coder/v2/site"
 	"github.com/coder/coder/v2/tailnet"
 	"github.com/coder/retry"
@@ -427,9 +427,9 @@ func (s *ServerTailnet) acquireTicket(agentID uuid.UUID) (release func()) {
 	}
 }
 
-func (s *ServerTailnet) AgentConn(ctx context.Context, agentID uuid.UUID) (*codersdk.WorkspaceAgentConn, func(), error) {
+func (s *ServerTailnet) AgentConn(ctx context.Context, agentID uuid.UUID) (*workspacesdk.WorkspaceAgentConn, func(), error) {
 	var (
-		conn *codersdk.WorkspaceAgentConn
+		conn *workspacesdk.WorkspaceAgentConn
 		ret  func()
 	)
 
@@ -440,9 +440,9 @@ func (s *ServerTailnet) AgentConn(ctx context.Context, agentID uuid.UUID) (*code
 	}
 	ret = s.acquireTicket(agentID)
 
-	conn = codersdk.NewWorkspaceAgentConn(s.conn, codersdk.WorkspaceAgentConnOptions{
+	conn = workspacesdk.NewWorkspaceAgentConn(s.conn, workspacesdk.WorkspaceAgentConnOptions{
 		AgentID:   agentID,
-		CloseFunc: func() error { return codersdk.ErrSkipClose },
+		CloseFunc: func() error { return workspacesdk.ErrSkipClose },
 	})
 
 	// Since we now have an open conn, be careful to close it if we error
