+++ enterprise/coderd/replicas_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/coder/coder/v2/coderd/coderdtest"
 	"github.com/coder/coder/v2/coderd/database/dbtestutil"
 	"github.com/coder/coder/v2/codersdk"
+	"github.com/coder/coder/v2/codersdk/workspacesdk"
 	"github.com/coder/coder/v2/enterprise/coderd/coderdenttest"
 	"github.com/coder/coder/v2/enterprise/coderd/license"
 	"github.com/coder/coder/v2/testutil"
@@ -82,10 +83,11 @@ func TestReplicas(t *testing.T) {
 		require.Len(t, replicas, 2)
 
 		r := setupWorkspaceAgent(t, firstClient, firstUser, 0)
-		conn, err := secondClient.DialWorkspaceAgent(context.Background(), r.sdkAgent.ID, &codersdk.DialWorkspaceAgentOptions{
-			BlockEndpoints: true,
-			Logger:         slogtest.Make(t, nil).Leveled(slog.LevelDebug),
-		})
+		conn, err := workspacesdk.NewWorkspaceClient(secondClient).
+			DialWorkspaceAgent(context.Background(), r.sdkAgent.ID, &workspacesdk.DialWorkspaceAgentOptions{
+				BlockEndpoints: true,
+				Logger:         slogtest.Make(t, nil).Leveled(slog.LevelDebug),
+			})
 		require.NoError(t, err)
 		require.Eventually(t, func() bool {
 			ctx, cancelFunc := context.WithTimeout(context.Background(), testutil.WaitShort)
@@ -128,10 +130,11 @@ func TestReplicas(t *testing.T) {
 		require.Len(t, replicas, 2)
 
 		r := setupWorkspaceAgent(t, firstClient, firstUser, 0)
-		conn, err := secondClient.DialWorkspaceAgent(context.Background(), r.sdkAgent.ID, &codersdk.DialWorkspaceAgentOptions{
-			BlockEndpoints: true,
-			Logger:         slogtest.Make(t, nil).Named("client").Leveled(slog.LevelDebug),
-		})
+		conn, err := workspacesdk.NewWorkspaceClient(secondClient).
+			DialWorkspaceAgent(context.Background(), r.sdkAgent.ID, &workspacesdk.DialWorkspaceAgentOptions{
+				BlockEndpoints: true,
+				Logger:         slogtest.Make(t, nil).Named("client").Leveled(slog.LevelDebug),
+			})
 		require.NoError(t, err)
 		require.Eventually(t, func() bool {
 			ctx, cancelFunc := context.WithTimeout(context.Background(), testutil.IntervalSlow)
