+++ enterprise/tailnet/pgcoord_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/coder/coder/v2/codersdk"
+	"github.com/coder/coder/v2/codersdk/workspacesdk"
 	agpltest "github.com/coder/coder/v2/tailnet/test"
 
 	"github.com/google/uuid"
@@ -228,7 +228,7 @@ func TestPGCoordinatorSingle_AgentValidIPLegacy(t *testing.T) {
 	defer agent.close()
 	agent.sendNode(&agpl.Node{
 		Addresses: []netip.Prefix{
-			netip.PrefixFrom(codersdk.WorkspaceAgentIP, 128),
+			netip.PrefixFrom(workspacesdk.WorkspaceAgentIP, 128),
 		},
 		PreferredDERP: 10,
 	})
