+++ enterprise/wsproxy/wsproxy_test.go
@@ -29,6 +29,7 @@ import (
 	"github.com/coder/coder/v2/coderd/httpmw"
 	"github.com/coder/coder/v2/coderd/workspaceapps/apptest"
 	"github.com/coder/coder/v2/codersdk"
+	"github.com/coder/coder/v2/codersdk/workspacesdk"
 	"github.com/coder/coder/v2/cryptorand"
 	"github.com/coder/coder/v2/enterprise/coderd/coderdenttest"
 	"github.com/coder/coder/v2/enterprise/coderd/license"
@@ -196,7 +197,7 @@ resourceLoop:
 		t.Parallel()
 
 		ctx := testutil.Context(t, testutil.WaitLong)
-		connInfo, err := client.WorkspaceAgentConnectionInfo(ctx, agentID)
+		connInfo, err := workspacesdk.NewWorkspaceClient(client).WorkspaceAgentConnectionInfo(ctx, agentID)
 		require.NoError(t, err)
 
 		// There should be three DERP regions in the map: the primary, and each
@@ -269,7 +270,7 @@ resourceLoop:
 		t.Parallel()
 
 		ctx := testutil.Context(t, testutil.WaitLong)
-		connInfo, err := client.WorkspaceAgentConnectionInfo(ctx, agentID)
+		connInfo, err := workspacesdk.NewWorkspaceClient(client).WorkspaceAgentConnectionInfo(ctx, agentID)
 		require.NoError(t, err)
 		require.NotNil(t, connInfo.DERPMap)
 		require.Len(t, connInfo.DERPMap.Regions, 3+len(api.DeploymentValues.DERP.Server.STUNAddresses.Value()))
@@ -429,13 +430,14 @@ resourceLoop:
 	_ = coderdtest.AwaitWorkspaceAgents(t, client, workspace.ID)
 
 	// Connect to the workspace agent.
-	conn, err := client.DialWorkspaceAgent(ctx, agentID, &codersdk.DialWorkspaceAgentOptions{
-		Logger: slogtest.Make(t, &slogtest.Options{
-			IgnoreErrors: true,
-		}).Named("client").Leveled(slog.LevelDebug),
-		// Force DERP.
-		BlockEndpoints: true,
-	})
+	conn, err := workspacesdk.NewWorkspaceClient(client).
+		DialWorkspaceAgent(ctx, agentID, &workspacesdk.DialWorkspaceAgentOptions{
+			Logger: slogtest.Make(t, &slogtest.Options{
+				IgnoreErrors: true,
+			}).Named("client").Leveled(slog.LevelDebug),
+			// Force DERP.
+			BlockEndpoints: true,
+		})
 	require.NoError(t, err)
 	t.Cleanup(func() {
 		err := conn.Close()
