+++ scaletest/reconnectingpty/run_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/coder/coder/v2/coderd/coderdtest"
 	"github.com/coder/coder/v2/coderd/httpapi"
 	"github.com/coder/coder/v2/codersdk"
+	"github.com/coder/coder/v2/codersdk/workspacesdk"
 	"github.com/coder/coder/v2/provisioner/echo"
 	"github.com/coder/coder/v2/provisionersdk/proto"
 	"github.com/coder/coder/v2/scaletest/reconnectingpty"
@@ -29,7 +30,7 @@ func Test_Runner(t *testing.T) {
 
 		runner := reconnectingpty.NewRunner(client, reconnectingpty.Config{
 			AgentID: agentID,
-			Init: codersdk.WorkspaceAgentReconnectingPTYInit{
+			Init: workspacesdk.WorkspaceAgentReconnectingPTYInit{
 				// Use ; here because it's powershell compatible (vs &&).
 				Command: "echo 'hello world'; sleep 1",
 			},
@@ -58,7 +59,7 @@ func Test_Runner(t *testing.T) {
 
 		runner := reconnectingpty.NewRunner(client, reconnectingpty.Config{
 			AgentID: agentID,
-			Init: codersdk.WorkspaceAgentReconnectingPTYInit{
+			Init: workspacesdk.WorkspaceAgentReconnectingPTYInit{
 				Command: "echo 'hello world'",
 			},
 			LogOutput: false,
@@ -86,7 +87,7 @@ func Test_Runner(t *testing.T) {
 
 			runner := reconnectingpty.NewRunner(client, reconnectingpty.Config{
 				AgentID: agentID,
-				Init: codersdk.WorkspaceAgentReconnectingPTYInit{
+				Init: workspacesdk.WorkspaceAgentReconnectingPTYInit{
 					Command: "echo 'hello world'",
 				},
 				Timeout:   httpapi.Duration(2 * testutil.WaitSuperLong),
@@ -110,7 +111,7 @@ func Test_Runner(t *testing.T) {
 
 			runner := reconnectingpty.NewRunner(client, reconnectingpty.Config{
 				AgentID: agentID,
-				Init: codersdk.WorkspaceAgentReconnectingPTYInit{
+				Init: workspacesdk.WorkspaceAgentReconnectingPTYInit{
 					Command: "sleep 120",
 				},
 				Timeout:   httpapi.Duration(2 * time.Second),
@@ -139,7 +140,7 @@ func Test_Runner(t *testing.T) {
 
 			runner := reconnectingpty.NewRunner(client, reconnectingpty.Config{
 				AgentID: agentID,
-				Init: codersdk.WorkspaceAgentReconnectingPTYInit{
+				Init: workspacesdk.WorkspaceAgentReconnectingPTYInit{
 					Command: "sleep 120",
 				},
 				Timeout:       httpapi.Duration(2 * time.Second),
@@ -164,7 +165,7 @@ func Test_Runner(t *testing.T) {
 
 			runner := reconnectingpty.NewRunner(client, reconnectingpty.Config{
 				AgentID: agentID,
-				Init: codersdk.WorkspaceAgentReconnectingPTYInit{
+				Init: workspacesdk.WorkspaceAgentReconnectingPTYInit{
 					Command: "echo 'hello world'",
 				},
 				Timeout:       httpapi.Duration(2 * testutil.WaitSuperLong),
@@ -194,7 +195,7 @@ func Test_Runner(t *testing.T) {
 
 			runner := reconnectingpty.NewRunner(client, reconnectingpty.Config{
 				AgentID: agentID,
-				Init: codersdk.WorkspaceAgentReconnectingPTYInit{
+				Init: workspacesdk.WorkspaceAgentReconnectingPTYInit{
 					Command: "echo 'hello world'; sleep 1",
 				},
 				ExpectOutput: "hello world",
@@ -218,7 +219,7 @@ func Test_Runner(t *testing.T) {
 
 			runner := reconnectingpty.NewRunner(client, reconnectingpty.Config{
 				AgentID: agentID,
-				Init: codersdk.WorkspaceAgentReconnectingPTYInit{
+				Init: workspacesdk.WorkspaceAgentReconnectingPTYInit{
 					Command: "echo 'hello world'; sleep 1",
 				},
 				ExpectOutput: "bello borld",
